util: remove shebang line
'util' is a source-able but not executable so the shebang line does not
make sense.

Also remove the "set -eu" to make it possible to test the functions in
an interactive session. All the callers set these options anyway.

Rename the file to util.bash so that syntax hilighting etc keeps
working.